Refactor/add mypy quality code

+++ app/ui/components/tropos_element_item/hard_goal_item.py
@@ -23,7 +23,7 @@ class HardGoalNodeItem(BaseTroposItem):
         super().__init__(HardGoal(x, y, radius))
 
     def _get_distance_to_border(self, pos: QPointF) -> float:
-        # ✅ Usar _independent_model si existe
+        # Usar _independent_model si existe
         model_for_props = (
             self._independent_model
             if hasattr(self, "_independent_model") and self._independent_model
@@ -36,22 +36,22 @@ class HardGoalNodeItem(BaseTroposItem):
             dist_right = abs(pos.x() - rect.right())
             dist_top = abs(pos.y() - rect.top())
             dist_bottom = abs(pos.y() - rect.bottom())
-            return min(dist_left, dist_right, dist_top, dist_bottom)
+            return float(min(dist_left, dist_right, dist_top, dist_bottom))
         else:
             dx = max(rect.left() - pos.x(), 0, pos.x() - rect.right())
             dy = max(rect.top() - pos.y(), 0, pos.y() - rect.bottom())
-            return math.sqrt(dx * dx + dy * dy)
+            return float(math.sqrt(dx * dx + dy * dy))
 
     def _get_new_radius_from_pos(self, pos: QPointF) -> float:
         new_r = abs(pos.x())
-        return max(new_r, 20.0)
+        return float(max(new_r, 20.0))
 
     def paint(self, painter, option, widget=None):
         default_color = QColor(150, 200, 150)
         default_border = QColor(0, 0, 0)
         default_text = QColor(255, 255, 255)
 
-        # ✅ Usar _independent_model si existe (para nodos composite internos)
+        # Usar _independent_model si existe (para nodos composite internos)
         model_for_props = (
             self._independent_model
             if hasattr(self, "_independent_model") and self._independent_model
@@ -77,7 +77,7 @@ class HardGoalNodeItem(BaseTroposItem):
         painter.setBrush(QBrush(fill_color))
         painter.setPen(QPen(border_color, 2))
 
-        # ✅ Usar radio del modelo independiente
+        # Usar radio del modelo independiente
         r = model_for_props.radius
         rect = QRectF(-r, -r / 2, 2 * r, r)
         path = QPainterPath()
